+++ src/config/config.dev.ts
@@ -1,0 +1,5 @@
+const configDev = {
+  modulePath: 'modules/'
+};
+
+window['AppConf'] = configDev;