+++ src/config/config.prod.ts
@@ -1,0 +1,5 @@
+const configProd = {
+  modulePath: 'modules/'
+};
+
+window['AppConf'] = configProd;