+++ src/global.d.ts
@@ -1,0 +1,3 @@
+declare const AppConf: {
+  modulePath: string
+};